(`c2rust-analyze`) Move `fn const_perms` to `pub fn PermissionSet::for_const`.
