+++ c2rust-analyze/src/dataflow/type_check.rs
@@ -5,10 +5,10 @@ use crate::util::{describe_rvalue, ty_callee, Callee, RvalueDesc};
 use assert_matches::assert_matches;
 use rustc_hir::def_id::DefId;
 use rustc_middle::mir::{
-    AggregateKind, BinOp, Body, ConstantKind, Location, Mutability, Operand, Place, PlaceRef,
-    ProjectionElem, Rvalue, Statement, StatementKind, Terminator, TerminatorKind,
+    AggregateKind, BinOp, Body, Location, Mutability, Operand, Place, PlaceRef, ProjectionElem,
+    Rvalue, Statement, StatementKind, Terminator, TerminatorKind,
 };
-use rustc_middle::ty::{self, SubstsRef, Ty, TyKind};
+use rustc_middle::ty::{SubstsRef, Ty, TyKind};
 
 /// Visitor that walks over the MIR, computing types of rvalues/operands/places and generating
 /// constraints as a side effect.
@@ -454,21 +454,6 @@ impl<'tcx> TypeChecker<'tcx, '_> {
     }
 }
 
-fn const_perms(constant: ConstantKind) -> PermissionSet {
-    let ref_ty = constant.ty();
-    let ty = match ref_ty.kind() {
-        ty::Ref(_, ty, _) => ty,
-        _ => panic!("expected only `Ref`s for constants: {ref_ty:?}"),
-    };
-    if ty.is_array() || ty.is_str() {
-        PermissionSet::READ | PermissionSet::OFFSET_ADD
-    } else if ty.is_primitive_ty() {
-        PermissionSet::READ
-    } else {
-        panic!("expected an array, str, or primitive type: {ty:?}");
-    }
-}
-
 pub fn visit<'tcx>(
     acx: &AnalysisCtxt<'_, 'tcx>,
     mir: &Body<'tcx>,
@@ -482,7 +467,7 @@ pub fn visit<'tcx>(
 
     for (&constant, const_lty) in &acx.const_ref_tys {
         tc.constraints
-            .add_all_perms(const_lty.label, const_perms(constant));
+            .add_all_perms(const_lty.label, PermissionSet::for_const(constant));
     }
 
     for (bb, bb_data) in mir.basic_blocks().iter_enumerated() {
@@ -506,7 +491,7 @@ pub fn visit<'tcx>(
 
     for (&constant, const_lty) in &acx.const_ref_tys {
         let _ptr_id = const_lty.label;
-        let _expected_perms = const_perms(constant);
+        let _expected_perms = PermissionSet::for_const(constant);
         // TODO: check that perms match the expected ones
     }
